feat: add simple logic to handle user logout
