+++ src/app/_actions/deleteUserToken.ts
@@ -1,0 +1,11 @@
+"use server"
+
+import { cookies } from "next/headers"
+import { redirect } from "next/navigation";
+
+
+export const deleteUserToken = async () => {
+    const cookieStore = await cookies();
+    cookieStore.delete("access_token")
+    redirect('/')
+}
