+++ src/app/_components/header.tsx
@@ -3,32 +3,13 @@ import Image from "next/image";
 import { Card } from "./ui/card";
 import { ToggleStyle } from "./ui/toggleStyle";
 import { UserAvatar } from "./ui/userAvatar";
-import { cookies } from "next/headers";
-import api from "@/utils/axios-instance";
-import { User } from "@/types/user";
-import { redirect } from "next/navigation";
+import { getUserData } from "../_actions/getUserData";
+import { deleteUserToken } from "../_actions/deleteUserToken";
 
 export const Header = async () => {
-  const fetchUserData = async (): Promise<User> => {
-    const cookieStore = await cookies();
-    const access_token = cookieStore.get("access_token");
-    if (!access_token) {
-      redirect("/");
-    }
-
-    try {
-      const response = await api.get("/google/me", {
-        headers: {
-          Authorization: `Bearer ${access_token.value}`,
-        },
-      });
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  };
-  const data = await fetchUserData();
-
+  const data = await getUserData();
+  
+  
   return (
     <header className="w-full bg-white">
       <Card className="flex justify-between items-center px-3 py-3">
@@ -43,7 +24,9 @@ export const Header = async () => {
         />
         <div className="flex items-center">
           <ToggleStyle />
-          <UserAvatar {...data} />
+          <button onClick={deleteUserToken} >
+            <UserAvatar {...data} />
+          </button>
         </div>
       </Card>
     </header>
